LPS-46842 No logic changes - move logic
